+++ Countly.m
@@ -1295,6 +1295,7 @@ static dispatch_once_t onceToken;
 
 - (void)attemptToSendStoredRequests
 {
+    CLY_LOG_I(@"%s", __FUNCTION__);
     [CountlyConnectionManager.sharedInstance attemptToSendStoredRequests];
 }