+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## X.x.x
 * Added a separate SDK Limits Config with the following options:
 ...
+* Fixed session duration inconsistency by incorporating checks for whether the session has started or not.
 
 ## 24.1.0
 * Added a separate APM Configs with following options: