+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 ## X.X.X
 * Added `attemptToSendStoredRequests` method to combine all events in event queue into a request and attempt to process stored requests
 * Added the iOS privacy manifest to the Countly SDK
+* Added a separate SDK Limits Config with the following options:
+    * `setMaxKeyLength`
+    * `setMaxValueSize`
+    * `setMaxBreadcrumbCount`
+    * `setMaxSegmentationValues`
+    * `setMaxStackTraceLineLength`
+    * `setMaxStackTraceLinesPerThread`
+    
 * Fixed session duration inconsistency by incorporating checks for whether the session has started or not.
+
+* Deprecated `maxKeyLength` initial config flag
+* Deprecated `crashLogLimit` initial config flag
+* Deprecated `maxValueLength` initial config flag
+* Deprecated `maxSegmentationValues` initial config flag
 
 ## 24.1.0
 * Added a separate APM Configs with following options: